Segregando em fila a importação do acervo fotográficos
