+++ SME.CDEP.Infra.Servicos/Rabbit/RotasRabbit.cs
@@ -24,5 +24,6 @@ namespace SME.CDEP.Infra
         public const string ExecutarImportacaoArquivoAcervoDocumental = "cdep.importacao.arquivo.acervo.documental";
         public const string ExecutarImportacaoArquivoAcervoArteGrafica = "cdep.importacao.arquivo.acervo.arte.grafica";
         public const string ExecutarImportacaoArquivoAcervoAudiovisual = "cdep.importacao.arquivo.acervo.audiovisual";
+        public const string ExecutarImportacaoArquivoAcervoFotografico = "cdep.importacao.arquivo.acervo.fotografico";
     }
 }
