+++ SME.CDEP.TesteIntegracao/ImportacaoArquivo/Ao_fazer_manutencao_importacao_arquivo_acervo_fotografico.cs
@@ -38,8 +38,8 @@ namespace SME.CDEP.TesteIntegracao
             acervoFotograficoLinhas[8].Codigo.Conteudo = string.Empty;
             var linhasComErros = new[] { 3, 5, 6, 8, 9 };
             
-            await servicoImportacaoArquivo.CarregarDominiosFotograficos();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosFotograficos();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
 
             foreach (var linha in acervoFotograficoLinhas)
             {
@@ -140,8 +140,8 @@ namespace SME.CDEP.TesteIntegracao
             acervoFotograficoLinhas[8].Credito.Conteudo = "Desconhecido 1 | Desconhecido 2";
             var linhasComErros = new[] { 3, 5, 6, 8, 9 };
             
-            await servicoImportacaoArquivo.CarregarDominiosFotograficos();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosFotograficos();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
 
             foreach (var linha in acervoFotograficoLinhas)
             {
@@ -247,8 +247,8 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosFotograficos();
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosFotograficos();
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoFotograficoLinhas);
         
             var acervos = ObterTodos<Acervo>();
             var acervosFotografico = ObterTodos<AcervoFotografico>();
@@ -337,9 +337,9 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosFotograficos();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosFotograficos();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoFotograficoLinhas);
             await servicoImportacaoArquivo.AtualizarImportacao(1, JsonConvert.SerializeObject(acervoFotograficoLinhas), acervoFotograficoLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
             var retorno = await servicoImportacaoArquivo.ObterImportacaoPendente();
         
@@ -486,9 +486,9 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosFotograficos();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosFotograficos();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoFotograficoLinhas);
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoFotograficoLinhas);
             await servicoImportacaoArquivo.AtualizarImportacao(1, JsonConvert.SerializeObject(acervoFotograficoLinhas), acervoFotograficoLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
             var retorno = await servicoImportacaoArquivo.ObterImportacaoPendente();
