+++ SME.CDEP.Worker/WorkerRabbitMQ.cs
@@ -46,6 +46,7 @@ namespace SME.CDEP.Worker
             Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoDocumental, new ComandoRabbit("Realiza a importação de acervos documental por planilha", typeof(IExecutarImportacaoArquivoAcervoDocumentalUseCase), true));
             Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoArteGrafica, new ComandoRabbit("Realiza a importação de acervos artes gráficas por planilha", typeof(IExecutarImportacaoArquivoAcervoArteGraficaUseCase), true));
             Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoAudiovisual, new ComandoRabbit("Realiza a importação de acervos audiovisual por planilha", typeof(IExecutarImportacaoArquivoAcervoAudiovisualUseCase), true));
+            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoFotografico, new ComandoRabbit("Realiza a importação de acervos fotográficos por planilha", typeof(IExecutarImportacaoArquivoAcervoFotograficoUseCase), true));
         }
     }
 }
